requirements.txt
absl-py==2.0.0
asgiref==3.7.2
astor==0.8.1
boto==2.49.0
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cycler==0.11.0
dataclasses==0.6
Django==3.2.24
fonttools==4.38.0
future==1.0.0
gast==0.2.2
google-pasta==0.2.0
grpcio==1.60.0
h5py==2.10.0
idna==3.6
importlib-metadata==6.7.0
imutils==0.5.4
joblib==1.3.2
Keras==2.2.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.5.3
mysqlclient==2.1.0
nltk==3.8.1
numpy==1.21.6
opencv-python==4.5.5.62
openpifpaf==0.13.11
opt-einsum==3.3.0
packaging==23.2
pandas==1.3.5
ParallelDots==3.2.14
Pillow==9.5.0
protobuf==3.19.3
pyparsing==3.1.1
pysparkling==0.6.2
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
scikit-learn==0.22.2.post1
scipy==1.7.3
seaborn==0.12.2
six==1.16.0
sqlparse==0.4.4
tensorboard==1.15.0
tensorflow-cpu==1.15.0
tensorflow-cpu-estimator==1.15.1
termcolor==2.3.0
threadpoolctl==3.1.0
torch==1.13.1
torchvision==0.14.1
tqdm==4.66.2
typing_extensions==4.7.1
urllib3==1.26.18
Werkzeug==2.2.3
wrapt==1.16.0
zipp==3.15.0
